moving everything into useLauncher
